[EI-252] [web] feat: Add metadata creation button to metadata list
